+++ README.md
@@ -12,7 +12,6 @@
 - [텍스트와 이미지로 살펴보는 온라인 코드 리뷰 과정](https://github.com/next-step/nextstep-docs/tree/master/codereview)
 
 ## 기능 목록
-<<<<<<< HEAD
 
 [x] 구입 금액을 입력받는다.
