+++ src/main/java/lotto/domain/LottoNumberGenerator.java
@@ -7,17 +7,20 @@ import java.util.stream.Collectors;
 import java.util.stream.IntStream;
 
 public class LottoNumberGenerator {
-    private static final List<Integer> SAMPLE = IntStream.rangeClosed(Lotto.MIN, Lotto.MAX)
+    public static final int START_NUMBER = 1;
+    public static final int END_NUMBER = 45;
+    public static final List<Integer> SAMPLE = IntStream.rangeClosed(START_NUMBER, END_NUMBER)
             .boxed()
             .collect(Collectors.toList());
 
     private LottoNumberGenerator() {
+
     }
 
     public static List<Integer> get() {
         List<Integer> shuffle = new ArrayList<>(SAMPLE);
         Collections.shuffle(shuffle);
-        shuffle = shuffle.subList(0, Lotto.SIZE);
+        shuffle = shuffle.subList(0, Lotto.LOTTO_SIZE);
         shuffle.sort(Integer::compareTo);
         return shuffle;
     }
