+++ src/test/java/lotto/domain/LottoFixture.java
@@ -1,5 +1,6 @@
 package lotto.domain;
 
+import lotto.domain.Lotto;
 import org.assertj.core.util.Lists;
 
 public class LottoFixture {
