+++ src/test/java/lotto/domain/LottoMachineTest.java
@@ -1,5 +1,6 @@
 package lotto.domain;
 
+import lotto.domain.LottoMachine;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
